+++ WordPress/Tests/NamingConventions/ValidTestClassName/ValidTestClassTest.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Test file for ValidTestClassName sniff.
+ */
+
+class ValidTestClassTest extends \PHPUnit_Framework_TestCase {
+
+    /**
+     * A test method.
+     */
+    public function testSomething() {
+        $this->assertTrue( true );
+    }
+}